Split HDFFile
