+++ src/coreComponents/fileIO/timeHistory/HDFHistoryIO.hpp
@@ -32,61 +32,6 @@ namespace geosx
 {
 
 /**
- * @class HDFFile
- * A class used to control access to an HDF file target.
- */
-class HDFFile
-{
-public:
-  /**
-   * @brief Constructor -- this creates/opens the target file for read/write.
-   * @param fnm The filename.
-   * @param deleteExisting Whether to remove/recreate if a file with the same name exists.
-   * @param parallelAccess Whether to access one file in parallel or one file per rank in the comm.
-   * @param comm An MPI communicator where each rank in the communicator will be accesing the target file.
-   */
-  HDFFile( string const & fnm, bool deleteExisting, bool parallelAccess, MPI_Comm comm );
-
-  /**
-   * Destructor -- Close the file and acccessors.
-   */
-  ~HDFFile();
-
-  /**
- * @brief Whether a dataset/group with the specified name exists in the target.
- * @param name The dataset/group name to check for.
- * @return Whether the dataset/group exists in the target.
- */
-  bool checkInTarget( const string & name )
-  {
-    htri_t exists = 0;
-    H5E_BEGIN_TRY
-      {
-        exists = H5Gget_objinfo( this->operator hid_t(), name.c_str(), 0, NULL );
-      }
-    H5E_END_TRY
-    return ( exists == 0 );
-  }
-
-  /**
-   * @brief Get the HDF hid_t file identifier.
-   * @return the HDF hid_t file id.
-   */
-  virtual operator hid_t() final { return m_fileId; }
-private:
-  /// The filename
-  string m_filename;
-  /// The hdf file id
-  hid_t m_fileId;
-  /// The hdf file access properties list id
-  hid_t m_faplId;
-  /// Whether to open the same file in parallel, or open a file per process
-  bool m_mpioFapl;
-  /// The comminator to operate on the file collectively over
-  MPI_Comm m_comm;
-};
-
-/**
  * @class HDFHistoryIO
  * @brief Perform buffered history I/O for a single type(really just output) on using HDF5.
  */
