+++ src/coreComponents/unitTests/fileIOTests/testHDFFile.cpp
@@ -1,7 +1,7 @@
-#include "fileIO/Outputs/TimeHistoryOutput.hpp"
 #include "fileIO/timeHistory/HDFHistoryIO.hpp"
-#include "dataRepository/BufferOpsDevice.hpp"
+#include "fileIO/timeHistory/HDFFile.hpp"
 #include "mainInterface/initialization.hpp"
+#include "dataRepository/BufferOpsDevice.hpp"
 
 #include <gtest/gtest.h>
